+++ src/task10/entry.py
@@ -1,4 +1,4 @@
-from src.models_handler.ollama_handler import generate_response_from_prompt
+from src.ollama_handler import generate_response_from_prompt
 from src.utils import load_sentences
 import os
 
@@ -12,7 +12,7 @@ def prompt_by_dataset(kind, sentence):
     return prompts.get(kind, None)
 
 
-def generate_reasoning(input_file, output_file, model_name, dataset_name):
+def inference_task_10(input_file, output_file, model_name, dataset_name):
     sentences = load_sentences(input_file)
     with open(output_file, "w") as outfile:
         for sentence in sentences:
@@ -43,13 +43,3 @@ def get_input_output_files(kind, dataset_name):
     os.makedirs(output_dir, exist_ok=True)
     output_file = f"{output_dir}/reasoning_{dataset_name}.txt"
     return input_file, output_file
-
-
-if __name__ == "__main__":
-    model_name = "llama2"
-    dataset_names = ["conceptnet", "atomic"]
-    for dataset_name in dataset_names:
-        input_file, output_file = get_input_output_files("base", dataset_name)
-        generate_reasoning(input_file, output_file, model_name, dataset_name)
-        input_file, output_file = get_input_output_files("nonsense", dataset_name)
-        generate_reasoning(input_file, output_file, model_name, dataset_name)
